I really tried to refactor it. LoL. Python folder management is crazy. Help Me!
