+++ april-music-player/assets/icons/configuration_icons/colors.ico
