+++ april-music-player/components/album_image_window.py
