+++ april-music-player/consts/HTML_LABELS.py
