+++ april-music-player/music_player/musicplayer.py
@@ -2,8 +2,8 @@ from PyQt6.QtGui import QIcon
 import os
 from PyQt6.QtCore import QThread, pyqtSignal, QObject
 from musicplayerworker import MusicPlayerWorker
-from easy_json import EasyJson
-from time import sleep
+from utils.easy_json import EasyJson
+
 
 class MusicPlayer(QObject):  # Inherit from QObject
     # Define the signal as a class attribute
