+++ april-music-player/plugins/notetaking.py
